+++ packages/react-i18n/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.1-alpha.1](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-i18n@0.2.1-alpha.0...@selfcommunity/react-i18n@0.2.1-alpha.1) (2022-08-04)
+
+
+### Bug Fixes
+
+* added semicolon for snippet notification (contribution type) ([4c1d736](https://github.com/selfcommunity/community-js/commit/4c1d736e2983c68855111dc78c06bba2a006b660))
+
+
+
+
+
+## [0.2.1-alpha.0](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-i18n@0.2.0...@selfcommunity/react-i18n@0.2.1-alpha.0) (2022-08-04)
+
+**Note:** Version bump only for package @selfcommunity/react-i18n
+
+
+
+
+
 # [0.2.0](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-i18n@0.2.0-alpha.2...@selfcommunity/react-i18n@0.2.0) (2022-08-04)
 
 **Note:** Version bump only for package @selfcommunity/react-i18n