+++ packages/react-ui/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.1-alpha.1](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-ui@0.3.1-alpha.0...@selfcommunity/react-ui@0.3.1-alpha.1) (2022-08-04)
+
+
+### Bug Fixes
+
+* added semicolon for snippet notification (contribution type) ([4c1d736](https://github.com/selfcommunity/community-js/commit/4c1d736e2983c68855111dc78c06bba2a006b660))
+
+
+
+
+
+## [0.3.1-alpha.0](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-ui@0.3.0...@selfcommunity/react-ui@0.3.1-alpha.0) (2022-08-04)
+
+
+### Bug Fixes
+
+* added link for snippet vie ([975870b](https://github.com/selfcommunity/community-js/commit/975870b575f57d4825098b0e36f7537147df34b8))
+* added link for snippet view ([d946334](https://github.com/selfcommunity/community-js/commit/d946334d0e0c3e644209aef872b965b23f575a3f))
+* Fix notification link ([812eb43](https://github.com/selfcommunity/community-js/commit/812eb4335bada9260f1c0e4dbe36e26fa827f48c))
+
+
+
+
+
 # [0.3.0](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-ui@0.3.0-alpha.4...@selfcommunity/react-ui@0.3.0) (2022-08-04)
 
 **Note:** Version bump only for package @selfcommunity/react-ui